MODIFY:리뷰 반영, ADD:searchBox 애니메이션 추가
