+++ src/components/Nav/Logo.scss
@@ -1,0 +1,17 @@
+.logo {
+  display: flex;
+  align-items: center;
+  font-size: 1.2rem;
+
+  .logoIcon {
+    width: 3rem;
+    height: 3rem;
+  }
+
+  .logoText {
+    margin-left: 1rem;
+    display: flex;
+    flex-direction: column;
+    font-weight: 600;
+  }
+}
